+++ srcs/reverse_rotate.c
@@ -6,7 +6,7 @@
 /*   By: ynakashi <ynakashi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/08 14:46:28 by ynakashi          #+#    #+#             */
-/*   Updated: 2022/02/11 20:55:21 by ynakashi         ###   ########.fr       */
+/*   Updated: 2022/02/12 09:06:20 by ynakashi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	reverse_rotate(t_stack *a)
 {
 	size_t	size;
-	int		i;
+	size_t	i;
 	int		*tmp;
 
 	if (!a)
